Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/user-timing/test_user_timing_mark_and_measure_exception_when_invoke_with_timing_attributes-expected.txt

CONSOLE MESSAGE: line 58: ReferenceError: Can't find variable: wp_test
Description

This test validates exception scenarios of invoking mark() and measure() with timing attributes as value.


FAIL exception test of performance.mark and performance.measure ReferenceError: Can't find variable: wp_test
